requirements.txt
Flask~=2.3.2
numpy~=1.24.3
opencv-python~=4.7.0.72
matplotlib~=3.7.1
Pillow~=9.5.0
WTForms~=3.0.1